fix: created page endpoint for /stac/catalog/{id}
