+++ sites/geohub/src/routes/(app)/management/stac/catalog/[id]/+page.ts
@@ -1,0 +1,19 @@
+import { StacCatalogs } from '$lib/config/AppConfig';
+import { error } from '@sveltejs/kit';
+import type { PageLoad } from './$types';
+
+export const load: PageLoad = async ({ params }) => {
+	const id = params.id;
+	const stac = StacCatalogs.find((s) => s.id === id);
+	if (!stac) {
+		throw error(404, `This stac ID (${id}) is not found.`);
+	}
+	const title = `${stac.name} | STAC Catalog management | GeoHub`;
+	const content = stac.name;
+
+	return {
+		title,
+		content,
+		stac
+	};
+};
